+++ lib/belated/exceptions.rb
@@ -1,0 +1,3 @@
+class Belated
+  class JobError < StandardError; end
+end