+++ lib/belated/job_wrapper.rb
@@ -53,7 +53,7 @@ class Belated
       self.retries += 1
       return if retries > max_retries
 
-      self.at = Time.now.utc + (retries.next**4)
+      self.at = Time.now + (retries.next**4)
       log "Job #{id} failed, retrying at #{at}"
       Belated.job_list.push(self)
     end